API.request() Raise exception if response is a 400-600
Using request's `#raise_for_status()` method to be sure that we don't
accidentally ignore bad responses.

It's a good thing to check the response status code before processing
it.

On the other end my guess is that a bad response would have ended up
with some `KeyError` or something similar when processing the response.
But it's still good to check and raise an exception as soon as possible.

request's documentation: https://requests.readthedocs.io/en/master/user/quickstart/#response-status-codes