Fix lookup of permission because of mispelled param
The `applicationId` was [accidentally renamed] to `application_id` (probably
muscle memory as snake_case is always used in Python).

The `applicationId` field is used to find the permission to view the WebApp
but because of this value had the wrong name it couldn't find it and the
resource was trying to re-create one but fail with a `400 BAD REQUEST`
because of the missing `applicationId`.

Auth0 Authorization API `POST /permissions` documentation:
https://auth0.com/docs/api/authorization-extension#create-permission

[accidentally renamed]: https://github.com/ministryofjustice/analytics-platform-concourse-auth0-client-resource/commit/9b7b0a9a0feb27e62aaf7efe9f307a1cba2d1789#diff-6298e12f3a46e8cb7acf80e544006988L61